Replace `sticker-*` by a single `sticker` command
- The old utilities are now subcommands of sticker.
- sticker has a `completions` subcommand that generates completion
  definitions for various shells.
- `sticker` and `sticker --version` report the crate's version.
- The various applications are made more uniform through the
  `StickerPipelineApp` and `StickerTrainApp` traits.
