+++ sticker-utils/src/subcommands/mod.rs
@@ -1,0 +1,20 @@
+mod dep2label;
+pub use dep2label::Dep2LabelApp;
+
+mod graph_metadata;
+pub use graph_metadata::GraphMetadataApp;
+
+mod prepare;
+pub use prepare::PrepareApp;
+
+mod pretrain;
+pub use pretrain::PretrainApp;
+
+mod server;
+pub use server::ServerApp;
+
+mod tag;
+pub use tag::TagApp;
+
+mod train;
+pub use train::TrainApp;
