+++ sticker-utils/src/subcommands/server.rs
@@ -3,12 +3,13 @@ use std::io::{BufReader, BufWriter, Write};
 use std::net::{TcpListener, TcpStream};
 use std::sync::Arc;
 
-use clap::Arg;
+use clap::{App, Arg, ArgMatches};
 use conllx::io::{ReadSentence, Reader, Writer};
 use stdinout::OrExit;
+use sticker_utils::{Config, Pipeline, SentProcessor, TaggerSpeed, TomlRead};
 use threadpool::ThreadPool;
 
-use sticker_utils::{app, Config, Pipeline, SentProcessor, TaggerSpeed, TomlRead};
+use crate::{StickerApp, StickerPipelineApp};
 
 static ADDR: &str = "ADDR";
 static THREADS: &str = "THREADS";
@@ -23,8 +24,26 @@ pub struct ServerApp {
 }
 
 impl ServerApp {
-    fn new() -> Self {
-        let matches = app::sticker_pipeline_app("sticker-server")
+    fn serve(&self, pipeline: Arc<Pipeline>, pool: ThreadPool, listener: TcpListener) {
+        for stream in listener.incoming() {
+            match stream {
+                Ok(stream) => {
+                    let app = self.clone();
+                    let pipeline = pipeline.clone();
+                    pool.execute(move || handle_client(app, pipeline, stream))
+                }
+                Err(err) => eprintln!("Error processing stream: {}", err),
+            }
+        }
+    }
+}
+
+impl StickerPipelineApp for ServerApp {}
+
+impl StickerApp for ServerApp {
+    fn app() -> App<'static, 'static> {
+        Self::pipeline_app("server")
+            .about("Annotation server")
             .arg(
                 Arg::with_name(THREADS)
                     .short("t")
@@ -39,15 +58,16 @@ impl ServerApp {
                     .help("Address to bind to (e.g. localhost:4000)")
                     .default_value("localhost:4000"),
             )
-            .get_matches();
+    }
 
+    fn parse(matches: &ArgMatches) -> Self {
         let batch_size = matches
-            .value_of(app::BATCH_SIZE)
+            .value_of(Self::BATCH_SIZE)
             .unwrap()
             .parse()
             .or_exit("Cannot parse batch size", 1);
         let configs = matches
-            .values_of(app::CONFIGS)
+            .values_of(Self::CONFIGS)
             .unwrap()
             .map(ToOwned::to_owned)
             .collect();
@@ -57,7 +77,7 @@ impl ServerApp {
             .map(|v| v.parse().or_exit("Cannot parse number of threads", 1))
             .unwrap();
         let read_ahead = matches
-            .value_of(app::READ_AHEAD)
+            .value_of(Self::READ_AHEAD)
             .unwrap()
             .parse()
             .or_exit("Cannot parse number of batches to read ahead", 1);
@@ -70,55 +90,34 @@ impl ServerApp {
             read_ahead,
         }
     }
-}
 
-impl Default for ServerApp {
-    fn default() -> Self {
-        Self::new()
-    }
-}
+    fn run(&self) {
+        let mut configs = Vec::with_capacity(self.configs.len());
+        for filename in &self.configs {
+            let config_file = File::open(filename)
+                .or_exit(format!("Cannot open configuration file '{}'", filename), 1);
+            let mut config =
+                Config::from_toml_read(config_file).or_exit("Cannot parse configuration", 1);
+            config
+                .relativize_paths(filename)
+                .or_exit("Cannot relativize paths in configuration", 1);
 
-fn main() {
-    let app = ServerApp::new();
+            // Parallel processing is useless without the same parallelism in Tensorflow.
+            config.model.inter_op_parallelism_threads = self.n_threads;
+            config.model.intra_op_parallelism_threads = self.n_threads;
 
-    let mut configs = Vec::with_capacity(app.configs.len());
-    for filename in &app.configs {
-        let config_file = File::open(filename)
-            .or_exit(format!("Cannot open configuration file '{}'", filename), 1);
-        let mut config =
-            Config::from_toml_read(config_file).or_exit("Cannot parse configuration", 1);
-        config
-            .relativize_paths(filename)
-            .or_exit("Cannot relativize paths in configuration", 1);
-
-        // Parallel processing is useless without the same parallelism in Tensorflow.
-        config.model.inter_op_parallelism_threads = app.n_threads;
-        config.model.intra_op_parallelism_threads = app.n_threads;
-
-        configs.push(config);
-    }
-
-    let pipeline =
-        Pipeline::new_from_configs(&configs).or_exit("Cannot construct tagging pipeline", 1);
-
-    let pool = ThreadPool::new(app.n_threads);
-
-    let listener =
-        TcpListener::bind(&app.addr).or_exit(format!("Cannot listen on '{}'", app.addr), 1);
-
-    serve(app, Arc::new(pipeline), pool, listener);
-}
-
-fn serve(app: ServerApp, pipeline: Arc<Pipeline>, pool: ThreadPool, listener: TcpListener) {
-    for stream in listener.incoming() {
-        match stream {
-            Ok(stream) => {
-                let app = app.clone();
-                let pipeline = pipeline.clone();
-                pool.execute(move || handle_client(app, pipeline, stream))
-            }
-            Err(err) => eprintln!("Error processing stream: {}", err),
+            configs.push(config);
         }
+
+        let pipeline =
+            Pipeline::new_from_configs(&configs).or_exit("Cannot construct tagging pipeline", 1);
+
+        let pool = ThreadPool::new(self.n_threads);
+
+        let listener =
+            TcpListener::bind(&self.addr).or_exit(format!("Cannot listen on '{}'", self.addr), 1);
+
+        self.serve(Arc::new(pipeline), pool, listener);
     }
 }
